Client refactor, fixes fast refresh (#379)
* Client refactor, fixes fast refresh

* eslint
