+++ src/viser/client/src/HoverContext.ts
@@ -1,0 +1,4 @@
+import React from "react";
+
+export const HoverableContext =
+  React.createContext<React.MutableRefObject<boolean> | null>(null);
