+++ src/viser/client/src/WebsocketFunctions.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import * as THREE from "three";
 import { Message } from "./WebsocketMessages";
-import { ViewerContext, ViewerContextContents } from "./App";
+import { ViewerContext, ViewerContextContents } from "./ViewerContext";
 
 /** Easier, hook version of makeThrottledMessageSender. */
 export function useThrottledMessageSender(throttleMilliseconds: number) {
